+++ test/surgex/parser/parsers/float_parser_test.exs
@@ -9,11 +9,14 @@ defmodule Surgex.Parser.FloatParserTest do
   test "valid input" do
     assert FloatParser.call("1") == {:ok, 1.0}
     assert FloatParser.call("1.5") == {:ok, 1.5}
+    assert FloatParser.call("1.5", min: 1.5, max: 1.5) == {:ok, 1.5}
   end
 
   test "invalid input" do
     assert FloatParser.call("1.3a") == {:error, :invalid_float}
     assert FloatParser.call("x") == {:error, :invalid_float}
     assert FloatParser.call("") == {:error, :invalid_float}
+    assert FloatParser.call("1.5", min: 1.6) == {:error, :out_of_range}
+    assert FloatParser.call("1.5", max: 1.4) == {:error, :out_of_range}
   end
 end