+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.14.0
+
+- Extended `Surgex.Parser.FloatParser` with `min` and `max` options
+
+## 2.13.0
+
+- Added `Surgex.Parser.ResourceParser`
+
 ## 2.12.1
 
 - Fixed `Surgex.RPC.Client` to support no services in the client